Moving getUser into middleware folder for better organization.
