+++ lib/controllers/logout.js
@@ -49,7 +49,7 @@ module.exports = function (req, res, next) {
   }
 
   function handleLogout(callback) {
-    helpers.getUser(req, res, function () {
+    middleware.getUser(req, res, function () {
       var account = req.user;
       cleanupSession(function () {
         handlePostLogout(account, callback);
