+++ lib/helpers/index.js
@@ -9,7 +9,6 @@ module.exports = {
   expandAccount: require('./expand-account'),
   getHost: require('./get-host'),
   getRequiredRegistrationFields: require('./get-required-registration-fields'),
-  getUser: require('./get-user'),
   getAppModuleVersion: require('./get-app-module-version'),
   handleAcceptRequest: require('./handle-accept-request'),
   writeJsonError: require('./write-json-error'),
