+++ lib/middleware/get-user.js
@@ -2,8 +2,8 @@
 
 var stormpath = require('stormpath');
 
-var createSession = require('./create-session');
-var expandAccount = require('./expand-account');
+var createSession = require('../helpers/create-session');
+var expandAccount = require('../helpers/expand-account');
 
 /**
  * This callback, when called, will simply continue processing the HTTP
