+++ lib/middleware/index.js
@@ -5,6 +5,7 @@ module.exports = {
   authenticationRequired: require('./authentication-required'),
   deleteCookies: require('./delete-cookies'),
   revokeTokens: require('./revoke-tokens'),
+  getUser: require('./get-user'),
   groupsRequired: require('./groups-required'),
   loginRequired: require('./login-required')
 };
