+++ lib/stormpath.js
@@ -108,7 +108,7 @@ module.exports.init = function (app, opts) {
   }
 
   function stormpathMiddleware(req, res, next) {
-    helpers.getUser(req, res, next);
+    middleware.getUser(req, res, next);
   }
 
   // Build routes.
@@ -162,7 +162,7 @@ module.exports.init = function (app, opts) {
       if (web.idSite.enabled) {
         addGetRoute(web.login.uri, controllers.idSiteRedirect({ path: web.idSite.loginUri }));
       } else {
-        router.use(web.login.uri, helpers.getUser);
+        router.use(web.login.uri, middleware.getUser);
         addGetRoute(web.login.uri, controllers.login);
         addPostRoute(web.login.uri, controllers.login);
       }
@@ -235,7 +235,7 @@ module.exports.init = function (app, opts) {
 
 function getUserMiddlewareProxy(nextMiddleware) {
   return function (req, res, next) {
-    helpers.getUser(req, res, nextMiddleware.bind(null, req, res, next));
+    middleware.getUser(req, res, nextMiddleware.bind(null, req, res, next));
   };
 }
 
@@ -251,7 +251,7 @@ module.exports.loginRequired = getUserMiddlewareProxy(middleware.loginRequired);
  *
  * @property getUser
  */
-module.exports.getUser = helpers.getUser;
+module.exports.getUser = middleware.getUser;
 
 /**
  * Expose the `groupsRequired` middleware.
